feat: organize class
