+++ cli.ini
@@ -11,10 +11,10 @@ componet=componet
 default=null
 
 [flags]
-autocompile=--autocompile
+autocompile=autocompile
 
 [flags-abbrv]
-autocompile=-autocomp
+autocompile=autocomp
 
 [helpers]
 description=teste CLI
