+++ src/data.py
@@ -1,0 +1,3 @@
+FLOW_PASSED = 0
+FLOW_DATA_INVALID = 1
+
