+++ src/main.py
@@ -3,8 +3,8 @@ import argparse
 import config
 from parse import Parse
 
-from Init import *
-from Compile import *
+import flows.Compile as compile
+import flows.Init as init
 
 parse:argparse.ArgumentParser
 
@@ -16,17 +16,17 @@ def main():
     parse.add_argument("token", nargs="?", default=config.data["tokens"]["default"], choices=config.data["tokens"].values(), help=config.data["helpers"]["tokens"])
     parse.add_argument("data", nargs="*", default="", help=config.data["helpers"]["data"])
 
-    parse.add_argument(config.data["flags"]["autocompile"], config.data["flags-abbrv"]["autocompile"], action="store_true", help=config.data["flagsHelpers"]["flagcompile"])
+    parse.add_argument(config.formatToFlag(config.data["flags"]["autocompile"], False), config.formatToFlag(config.data["flags-abbrv"]["autocompile"], True), action="store_true", help=config.data["flagsHelpers"]["flagcompile"])
 
     args = parse.parse_args()
     parseArgs(args)
 
 def parseArgs(args):
     global parse
-    p = Init(args, parse)
+    p = init.Init(args, parse)
     Parse(p)
 
-    p2 = Compile(args, parse)
+    p2 = compile.Compile(args, parse)
     Parse(p2)
 
 if __name__ == "__main__":
